Add game assets and implement App class for game management
- Added ghost and pacman movement sprites.
- Created App class to manage game initialization and rendering.
- Updated GlobalRenderer and GameRenderer to handle player direction and input events.
- Refactored Actor and Player classes to support float coordinates.

+++ src/app/game/Actor.py
@@ -1,8 +1,8 @@
 class Actor:
     def __init__(
         self,
-        x: int = 0,
-        y: int = 0
+        x: float = 0.0,
+        y: float = 0.0
     ) -> None:
         self.x = x
         self.y = y
